Update to cmake 3.18
One of the problems was that CUDA CMAKE_CUDA_STANDARD in versions before
3.17 overrides the CMAKE_CXX_STANDARD. Now that I look at it, it was not
a problem for us but I updated anyway.

Once we updated to cmake above 3.18, I switched to using the newer way
of including CUDA in cmake, using the CUDAToolkit target instead of
deprecated CUDA target.

This also allows us to use the CMAKE_CUDA_ARCHITECTURES instead of
having to pass it as compiler options.

Enumerate the files to be compiled explicitly instead of globbing them.
This allows automatic cmake rebuild when we add new source file as the
CMakeLists.txt file will be changed.

Fix the target_compile_options, which we were manually commenting and
uncommenting when debuging/benchmarking. Now they are assigned to
Debug/Release builds.