Add row distribution
Add first two implementations of row distribution.

The rectangle distribution runs the maximum number of used workers by
any shift for all shifts and filters out the unnecessary workers. This
is simple to compute, but adds a lot of overhead by starting twice the
amount of threads needed.

The triangle distribution is much more complex. It creates a triangle of
threads, which has the growth rate as argument. For 3 max rows per
thread and 10x10 inputs, we have 19 rows of output with
1,2,3,4,5,6,7,8,9,10,9,8,7,6,5,4,3,2,1 rows which need to be processed
for each output row.

We assign the workers from top of the triangle, i.e. the worker with id
0 is the leftmost item on the top row of the triangle.

The triangle itself has 1 item height above 1,2,3, 2 item height above
4,5,6, three items height above 7,8,9 and 4 items height above 10.
The workers are sequential from left to right in each row, from top to
bottom between rows. So 0 is the top row, then 1 2 3 4 5 6 7 is the
second row, 8 9 10 11 12 13 14 15 16 17 18 19 20 the third row etc.

We can compute the ID of the leftmost worker in each row by
max_rows_per_thread = r, workers_on_top_row = t, leftmost_worker_in_row
= i, row = x (indexed from 1): r*x^2 + (3r - t)*x + (2r - t) = i. This
is basically expanded from
https://stackoverflow.com/questions/37513699/find-row-of-pyramid-based-on-index
,where I tried to interpolate polynomials through these triangles for
different sizes and max_rows_per_thread values and this solution worked
for all of them.

If we solve this for x, and we take the possitive root, we get floor((3r
- t) + sqrt((3r - t)^2 - 4r(2r - t - i))/(2r), which is basically using
the discriminant and the (-b +- sqrt(D))/2a. This allows us to compute
the output row (which corresponds to the computed shift), the number of
workers for given output row and the idx of the worker from workers for
this row.

The current implementation works only for search_size == 2*matrix_size -
1, as it requires the triangle to be complete. It can be upgraded to
handle even truncated triangles.

Currently it seems that the triangle distribution is too expensive to
compute compared to the amount of work done by each worker, mostly due
to the square root and the amount of integer division/modulo and
multiplication.

For the naive_Warp_per_shift algorithm, it is not really needed as the
algorithm achieves high occupancy even without this change. Future
commits will add the distribution to other algorithms which may benefit
more.

Change benchmark name generation to always append the arguments part
even if arguments are empty, so that we can parse the name easily and
separate the name from arguments.

Change main.cpp to CUDA language in CMAKE, so that it is passed through
nvcc before getting compiled by gcc. Before it was getting passed
straight to gcc, so I have no idea how the kernel calls were working at
all.

Change algorithms map in the main.cpp file from templated static
variable to templated functino returning a map. The previous
implementation was crashing nvcc once we started compiling it with nvcc
to get the proper compilation of kernel calls and device functions.
It was basically this error, which is originally from 2008. Some problem
with templates most likely.
https://forums.developer.nvidia.com/t/cuda-infernal-error-assertion-failed-gen-initializer-consta/5993