Add warp per shift with shared memory ring buffer
Add implementation of warp per shift with added shared memory.
The shared memory usage is derived from the warp_shuffle algorithm, but
instead of shuffling in x axis, we shuffle along the y axis.

We again have ring buffer, this time of two submatrices, reading from
left matrix, and a single submatrix from right matrix.

We preload the bottom half of the left buffer, prefixing it with zeroed
rows based on the block min shift. This is the same as warp_shuffle
values loaded by the first few threads being out of bounds before the
left matrix (i.e. negative index) and being loaded as zero. If we did
not do this, some of the rows in the left bottom buffer would correspond
to rows from the right matrix which will be loaded into the buffer in
second iteration, where the bottom part of the left buffer will already
be discarded. We need to only load left rows which have a corresponding
right row in the first rows loaded to the right buffer. This is done by
taking the left_buffer_start_row and if it starts as negative, prefixing
the loaded values with abs(left_buffer_start_row) zeroed out rows. This
way we only load the required left rows.

The kernel processes the matrices in cutouts of shared_mem_row_size
columns, i.e. it goes from top to bottom, loading the first
shared_mem_row_size values from each row for the given shift computed by
the warps of the block. Warps of each block compute the same x shift,
but different y shifts. THis allows us to load reuse rows loaded to
shared matrix for computation of different shifts by different warps.

We do it in the y axis because doing it in the x axis, as can be seen in
the (for now) commented out implementation
ccn_shift_per_warp_shared_mem, results in strided accesses to shared
memory, which lead to bank conflicts. Because we go row by row, the rows
read by each warp are continuous in the shared memory, resulting in nice
coalesced access without any bank conflicts. The only difference between
warps of each block is the offset they start reading at from shared
memory. This offset is in multiple of rows, as they may skip a first few
and/or last few rows in the column cutout of the matrix.

After the column cutout of the matrix is processed, the whole block
continues with the next cutout of the matrix of maximum size
shared_mem_row_size * matrix_size.y, until it goes through all the parts
of the matrix required to compute the shifts computed by warps of the
given block.

In the future, we may add work distribution between warps where multiple
warps may compute the same shift. This can be done by spliting the
computed rows of each warp, or by splitting the column cutouts of the
matrix between different warps, or both.

We may also try optimization for one_to_many where we will load multiple
right matrix cutouts for each left matrix cutout, so that we reuse the
data loaded to the left buffer for computation of multiple right
matrices.

The implementation currently contains debug prints as it still does not
work for matrices larger than 10x10, and to preserve them as examples of
how we did debugging.

Update the shared_mem.cuh implementation so that it can load column
cutout of a matrix. It reads chunks, in our case of size
shared_mem_row_size, with matrix_size.x - shared_mem_row_size stride
between chunks. But this implementation can be used to read any data
with consistently spaced chunks of a given size.

Add move constructor and swap implementation to shared_mem_buffer so
that we can implement ring buffer by swapping the whole buffers.

Add explicit public methods for load_continuous_chunk implementations.
load_continuous loads a continuous data from memory, either using
strided warps, where each warp loads warp_size of data each time offset
by thread_block.num_threads() values. This is the
load_continuous_chunk_strided_warps. The alternative is to split the
whole chunk to be loaded between the warps so that each warp receives a
continuous chunk of data to be loaded. This is the
load_continuous_chunk_continuous_warps.

The load_strided_chunk method also has these two implementations.
INstead of offset in the source data, the offset between data loaded by
strided_warps is now in the shared_memory it writes to. The
continuous_warps write into continuous part of shared memory. The other
difference is in small loads, where with continuous warps each warp
receives small amount of data to load. For strided_warps, the minimal
amount of data each warp loads is warp_size, so most warps will not load
anything. WE SHOULD TEST AND MEASURE THIS.

Remove redundant includes of shared_mem.cuh in some implementations.