Add check for shared_mem_rows < shifts_per_block. Measure warp pere shift algorithm
Current implementation does not work with shared_mem_rows <
shifts_per_block, so check the parameters at alg construction time and
throw the new argument_error, which is used to signal to the user that
an arugment for the given algorithm was given an invalid value.

Add seaborn to visualization.

Add name, args and input size as columns to the pandas dataframe in
visualization.

Make benchmark groups a dictionary instead of list, so that we can query
groups by name instead of index.