Add Visualization of warp shuffle one_to_many results
Add visualization of warp shuffle one to many results. From the
measurements it seems that for larger number of right matrices, as the
ration of shuffles to arithmetic operations falls, the algorithm becomes
better and better compared to the original naive. We are up to 70%
faster than the original naive, which is nice.

Rename few methods in shared.py InputSize and Run, to better correspond
to what they return. Before the methods input_area were not actually
returning area of the whole input, just the area of one matrix, so now
the method is correctly named matrix_area.

Add method total_items, which now returns total number of items in the
whole input.

Add equality implementation to InputSize, so that we don't have to
compare string representations.

Add columns to the pandas dataFrame so that we can see clearly the
results when printing.