Add rerun and continue modes to benchmarking
Instead of unconditionally skipping any group that already has the
existing results directory, we now allow users to choose whether the
whole script should fail (default), delete the previous results and
rerun the measurements or continue from where the last run left off.

This is mainly targeted for future measurements on gpulab cluster where
we could be interrupted. The whole rewrite is thus aimed at allowing
recovery from interrupt and the ability to continue when restarted.

Care is taken to make things interruptible and prevent data corruption.
Any file generation, be it validation data or measurement results, is
now done into a temp file first, which is then moved/renamed into the
target file. This should be atomic, so we don't have to worry about data
corruption.

This commit also contains a greater rewrite of the benchamrking
implementation, as the single run method was starting to get pretty
long. We now represent the run stages using classes. GroupExecution
holds the state specific for the execution of the group. Group itself
holds constant configuration loaded from file.

InputSizeSubgroup handles the subgroup of runs which work with given
inputSize, handling input data generation, validation data generation
and running the benchmark itself.

Changed the keep flag to keep_output, as we will be keeping inputs
unconditionally. Inputs are small and it simplifies the implementation
of continuation.

Added Logger class which centralizes the implementaion of logging. It is
not a nice design but it helps.

Add clear command to benchmarking which can be used to clear whole
benchmark suite or specific groups out of a suite. Just to simplify our
life from constant directory removal.