Pass CUDA cooperative group objects as const reference
The CUDA c++ Programming guide suggests to pass the cooperative groups
using const reference: "and pass these group objects by reference to
functions that intend to use these threads in some cooperative fashion",
which is also done in all their examples. This commit changes the
implmentation to follow that suggestion.