Upgrade time measurements, add outer loop measuring compute only
CPU and even GPU measurements which measure time shorter than few
milliseconds, which is the case when measuring a single run of a kernel
with smaller input sizes, are imprecise due to thread scheduling etc.

This change adds the ability to run the given block of code multiple
times and measure the total runtime. It also adds tracking of the number
of iterations throughout the code into the output so that we can work
with it when processing the results.

Instead of CPU_MEASURE and CUDA_MEASURE macros, we now have multiple
different forms. CPU_MEASURE is preserved and still does what it used
to, measuring a single run of the provided block of code. Additionally
we now have CPU_REPEATED_MEASURE, which runs the code given number of
times and measures that, and CPU_ADAPTIVE_MEASURE, which first runs the
code once, measures it and if required iteratively increases the number
of iterations until the code runs for at least the given time.

For CUDA_MEASURE, we instead have CUDA_SYNC_MEASURE, which synchronizes
on the events and reads the time immediately, CUDA_ASYNC_MEASURE, which
just inserts the cuda events which are then collected at the end, then
REPEATED versions of both, and finally the CUDA_ADAPTIVE_MEASURE, which
is basically CUDA_SYNC_REPEATED_MEASURE which adaptively changes the
number of iterations to run at least the given time.

The minimum time to run for ADAPTIVE versions is a property of the
stopwatch.

Changed the macros so that they don't expect anything from the context
they are used in. Instead they receive everything through arguments.
This allows the macros to be used anywhere, even outside the
cross_corr_alg classes.

Move the cudaDeviceSynchronize, which was scattered throughout the
implementation methods, to the measurement macros. Now each CPU macro
takes bool parameter if it should synchronize CUDA device when
measuring, to measure anything running on the device.

Split helpers.cuh file into host_helpers and cuda_helpers. Mainly we
split off the cuda_helpers to better represent what the file contains.

Add outer loop for measuring compute stages only. This allows us to load
the data once, which is very slow on the gpuLab cluster, and run the
computation multiple times. We can thus measure many more iterations in
the same amount of time. The number of loops is passed as a commandline
argument to the cross executable.

Add Benchmark types, which allows us to select the level we are
measuring and disable all other levels, so that they do not interfere
with each other. This is required for adaptive and repeated
measurements, as if the measured block of code contains another ADAPTIVE
or REPEATED measurement, the measured time is nonsensical. We caurrently
define three levels of measurement. We ahve Computation, which measures
all the computation stages as one, basically Total from previous
implementation without the loading stage. This is what we were mostly
working in visualization, so we want the measurement to be as simple as
possible. This measurement disables all measurements in the
cross_corr_alg base class and in each algorithm class.

The second level is CommonSteps, which measures each of the common steps
defined by the corss_corr_alg base class. This disables the other two
levels.

The last level is Algorithm, which measures algorithm specific parts of
code. This again disables the other two levels.

Lastly there is the None level, which disables all measurements. This
may be used for validation of correctness.

Merge the arguments for the run command into a single class, which pulls
it out of boost::program_options variable map and owns the values. This
class is then passed through the dispatch methods which take runtime
values of the commandline arguments and dispatch the execution to a
compile time templated method based on the value.

CHange the start and stop methods of cross_corr_alg base class into
reset_measurements and collect_measurements, which better represent what
these methods do.

CHange handling of measurement labels. Now they are not merged, only
selected based on the benchmark type.