Implement multirow shuffle
The new multirow shuffle algorithm tries to reduce the ration between
shuffles and arithmetic operations by shuffling multiple rows from a
single right matrix at once. This is very similar to the other shuffle
algorithms, which shuffle rows from multiple right matrices or even from
multiple left and right matrices.

The main difference is that in previous algorithms, each thread computed
a single shift, just accross many matrices. This made the code very
easy, as all bounds and limits were the same in all matrices, we just
repeated the same operation across many data inputs.

THe new algorithm is much harder, both to understand and implement, as
now each thread computes multiple different shifts, where each of the
shifts represents different matrix overlap and thus results in different
bounds in the rows processed by the shift.

As we are shuffling in a row, the only reasonable choice was to shuffle
multiple consecutive rows at once. This means that each thread computes
multiple shifts, where the matrices are shifted by [x,y],[x,y+1],[x,y+2]
up to a limit.

The main difference is that we need to run a startup phase and wind-down
phase, where we compute the few initial and final rows where not all of
the shifts overlap. We have the warp_right_start.y, which tells us the
first row in the right matrix which overlaps in at least one shift. The
exact row will overlap in exactly one shift, and each subsequent row
will overlap in one more shift up to the number of shifts. Inversely, we
have warp_right_end.y, which is a row AFTER the last row that overlaps
in any of the shifts. Again, the decrease is one per row down to the end
-1 row.

If the shifts do not overlap in the first few rows, the few steps of
startup are skipped. Same if the shifts do not overlap in the last few
rows, the last few steps of wind-down are skipped.

During startup, we first load a single row and compute it with the
corresponding row from the left matrix. The next iteration loads two
rows from the right matrix, the warp_right_start.y as in the first
iteration and warp_right_start.y + 1. We also need to load the next row
from left, which corresponds to min_shift + 1. The calculation then has
the min_shift result in the second place of the sum array and the result
of min_shift + 1 in the first place of the sum array. We have to invert
the order and add these row sums in reverse, so res[0] += sum[1] and
res[1] += sum[0].

It would be very hard, if not impossible, to keep the results in
registers across the device function boundaries. Thus we store the
results in shared memory, where each thread of the block accumulates the
result for the shifts it is computing. We need <num_threads> *
<num_shifts> elements of shared memory, which is rather low amount so it
should be fine.

In the shared memory, we first have the first shift of all threads of
the block, then second shift of all threads etc. So the shared memory
contains 256 (for example) elements for first shifts, then 256 elements
for second shifts etc. The max number of shifts per thread is currently
8.

As described above, the sum array which we use to sum each row contains
results in reverse order, having the last shift at index 0, and the
first shift at last place.

During wind-down, only the last few shifts overlap. Due to this we have
to pass the res array shifted to exclude the first few shifts which no
longer overlap, so we can add the row sums to the correct results.

Fix invalid type of the min shift in the original shuffle and n_to_m
shuffle args structure. Shifts may be negative, so it has to be an int,
but it was stored as unsigned int.

Add simple tool for debugging, which computes the expected row sums for
given parts of the two matrices.