Fix benchmarking scripts for validation
Fix benchmarking with "None" benchmark type useful for validation.

Fix result stats validation crashing when the benchmark fails. Now we
instead catch the expcetion which is caused by the empty CSV file and
display it as ERROR entry in the output.

Fix validation result generation. THe adaptive limit argument added to
matlab benchmarking was not added to the valid result computation. This
commit adds the argument with value 0, which causes matlab to run single
iteration only.