Add initial implementation of multirow_multiright optimization
Add initial implementation of the multirow_multiright optimization,
which combines the multirow optimization with optimization across
multiple right matrices.

Currently not working, needs to be fixed in the future. Some indexing
problem or something.

The implementation is pretty straight forward. From the base multirow
code, we just added one template argument, increased the size of sum
array and thread_right array and changed the single for cycle into two
for cycles, one across the shifts and one across the matrices.

TODO: In the future, we should try to change the nested for loops into a
single for loop. It may fix the problem with nvcc pulling the arrays
into local memory.